Draft implementation

+++ back/app/models/static_page.rb
@@ -28,11 +28,17 @@
 #  bottom_info_section_enabled  :boolean          default(FALSE), not null
 #  bottom_info_section_multiloc :jsonb            not null
 #  header_bg                    :string
+#  space_id                     :uuid
 #
 # Indexes
 #
-#  index_static_pages_on_code  (code)
-#  index_static_pages_on_slug  (slug) UNIQUE
+#  index_static_pages_on_code      (code)
+#  index_static_pages_on_slug      (slug) UNIQUE
+#  index_static_pages_on_space_id  (space_id)
+#
+# Foreign Keys
+#
+#  fk_rails_...  (space_id => spaces.id)
 #
 class StaticPage < ApplicationRecord
   include Files::FileAttachable
@@ -42,7 +48,7 @@ class StaticPage < ApplicationRecord
 
   slug from: proc { |page| page.title_multiloc&.values&.find(&:present?) }, except: RESERVED_SLUGS
 
-  enum :projects_filter_type, { no_filter: 'no_filter', areas: 'areas', global_topics: 'topics' }
+  enum :projects_filter_type, { no_filter: 'no_filter', areas: 'areas', global_topics: 'topics', spaces: 'spaces' }
 
   has_many_text_images from: :top_info_section_multiloc, as: :top_info_section_text_images
   has_many_text_images from: :bottom_info_section_multiloc, as: :bottom_info_section_text_images
@@ -56,6 +62,8 @@ class StaticPage < ApplicationRecord
 
   has_many :areas_static_pages, dependent: :destroy
   has_many :areas, through: :areas_static_pages
+
+  belongs_to :space, optional: true
 
   accepts_nested_attributes_for :nav_bar_item
 
@@ -118,6 +126,7 @@ class StaticPage < ApplicationRecord
       global_topics? && !Current.loading_tenant_template
     end
   )
+  validates :space, presence: true, if: -> { spaces? && !Current.loading_tenant_template }
 
   mount_base64_uploader :header_bg, HeaderBgUploader
 
@@ -138,6 +147,8 @@ class StaticPage < ApplicationRecord
         { areas: areas_static_pages.pluck(:area_id) }
       when 'global_topics'
         { global_topics: static_pages_global_topics.pluck(:global_topic_id) }
+      when 'spaces'
+        { spaces: [space_id].compact }
       else
         {}
       end
@@ -199,7 +210,12 @@ class StaticPage < ApplicationRecord
     # projects_filter_type returns the key as a string, and .keys also returns strings
     current_key = projects_filter_type
     (self.class.associations_project_filter_types.keys - [current_key]).each do |association|
-      public_send(association).destroy_all
+      case association
+      when 'spaces'
+        self.space_id = nil
+      else
+        public_send(association).destroy_all
+      end
     end
   end
 end
